Feat: Adds a stable version for the filter and sorting
